+++ core/include/booker/utils/StringUtils.hpp
@@ -1,0 +1,13 @@
+#ifndef STRINGUTILS_HPP
+#define STRINGUTILS_HPP
+
+/*!
+\file StringUtils.hpp
+\brief Define utils functions related to strings.
+*/
+
+#include <string>
+
+std::string fillWithLeadingZeros(int nb, int maximum);
+
+#endif // STRINGUTILS_HPP